requirements.txt
absl-py==2.3.1
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bitsandbytes==0.47.0
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.2
contourpy==1.3.3
cycler==0.12.1
datasets==4.0.0
debugpy==1.8.11
decorator==5.1.1
deepspeed==0.17.4
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.17.0
einops==0.8.1
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.18.0
flash_attn==2.6.3
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.16.0
hf-xet==1.1.7
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
immutables==0.21
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
lark==1.2.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nltk==3.9.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.102.0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
peft==0.17.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.6
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.48
propcache==0.3.2
protobuf==5.29.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rouge_score==0.1.2
rpds-py==0.27.0
ruff==0.12.8
safetensors==0.6.2
Send2Trash==1.8.3
sentencepiece==0.2.1
sentry-sdk==2.19.2
setproctitle==1.3.4
shtab==1.7.2
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
-e git+ssh://git@github.com/DIR-LAB/streamingcheck.git@5b0ca63abd4f59e9d1c4da807bb6bd0866d8107b#egg=streamingcheck
sympy==1.14.0
terminado==0.18.1
tinycss2==1.4.0
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.0
triton==3.4.0
trl==0.21.0
typeguard==4.4.4
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
tyro==0.9.31
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wandb==0.19.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xxhash==3.5.0
yarl==1.20.1
